Separated user profiles from user credentials and renamed Socials context to Publications
