+++ lib/vutuv/accounts/user_credential.ex
@@ -4,7 +4,7 @@ defmodule Vutuv.Accounts.UserCredential do
   import Ecto.Changeset
 
   alias NotQwerty123.PasswordStrength
-  alias Vutuv.Accounts.User
+  alias Vutuv.UserProfiles.User
 
   @type t :: %__MODULE__{
           id: integer,
