+++ lib/vutuv/user_profiles/address.ex
@@ -1,9 +1,9 @@
-defmodule Vutuv.Accounts.Address do
+defmodule Vutuv.UserProfiles.Address do
   use Ecto.Schema
 
   import Ecto.Changeset
 
-  alias Vutuv.Accounts.User
+  alias Vutuv.UserProfiles.User
 
   @type t :: %__MODULE__{
           id: integer,
