+++ lib/vutuv_web/auth/login.ex
@@ -6,7 +6,7 @@ defmodule VutuvWeb.Auth.Login do
 
   use Phauxth.Login.Base
 
-  alias Vutuv.Accounts
+  alias Vutuv.{Accounts, UserProfiles}
 
   @impl true
   def authenticate(%{"password" => password} = params, _, opts) do
@@ -18,7 +18,7 @@ defmodule VutuvWeb.Auth.Login do
   end
 
   defp get_user_struct({:ok, %{user_id: user_id}}) do
-    {:ok, Accounts.get_user(user_id)}
+    {:ok, UserProfiles.get_user(user_id)}
   end
 
   defp get_user_struct({:error, message}), do: {:error, message}
