+++ lib/vutuv_web/controllers/address_controller.ex
@@ -3,8 +3,8 @@ defmodule VutuvWeb.AddressController do
 
   import VutuvWeb.Authorize
 
-  alias Vutuv.Accounts
-  alias Vutuv.Accounts.{Address, User}
+  alias Vutuv.UserProfiles
+  alias Vutuv.UserProfiles.{Address, User}
 
   @dialyzer {:nowarn_function, new: 3}
 
@@ -17,23 +17,23 @@ defmodule VutuvWeb.AddressController do
   end
 
   def index(conn, %{"user_slug" => slug}, %User{slug: slug} = current_user) do
-    addresses = Accounts.list_addresses(current_user)
+    addresses = UserProfiles.list_addresses(current_user)
     render(conn, "index.html", addresses: addresses, user: current_user)
   end
 
   def index(conn, %{"user_slug" => slug}, _current_user) do
-    user = Accounts.get_user!(%{"slug" => slug})
-    addresses = Accounts.list_addresses(user)
+    user = UserProfiles.get_user!(%{"slug" => slug})
+    addresses = UserProfiles.list_addresses(user)
     render(conn, "index.html", addresses: addresses, user: user)
   end
 
   def new(conn, _params, _current_user) do
-    changeset = Accounts.change_address(%Address{})
+    changeset = UserProfiles.change_address(%Address{})
     render(conn, "new.html", changeset: changeset)
   end
 
   def create(conn, %{"address" => address_params}, current_user) do
-    case Accounts.create_address(current_user, address_params) do
+    case UserProfiles.create_address(current_user, address_params) do
       {:ok, address} ->
         conn
         |> put_flash(:info, gettext("Address created successfully."))
@@ -45,26 +45,26 @@ defmodule VutuvWeb.AddressController do
   end
 
   def show(conn, %{"id" => id, "user_slug" => slug}, %User{slug: slug} = current_user) do
-    address = Accounts.get_address!(current_user, id)
+    address = UserProfiles.get_address!(current_user, id)
     render(conn, "show.html", address: address, user: current_user)
   end
 
   def show(conn, %{"id" => id, "user_slug" => slug}, _current_user) do
-    user = Accounts.get_user!(%{"slug" => slug})
-    address = Accounts.get_address!(user, id)
+    user = UserProfiles.get_user!(%{"slug" => slug})
+    address = UserProfiles.get_address!(user, id)
     render(conn, "show.html", address: address, user: user)
   end
 
   def edit(conn, %{"id" => id}, current_user) do
-    address = Accounts.get_address!(current_user, id)
-    changeset = Accounts.change_address(address)
+    address = UserProfiles.get_address!(current_user, id)
+    changeset = UserProfiles.change_address(address)
     render(conn, "edit.html", address: address, changeset: changeset)
   end
 
   def update(conn, %{"id" => id, "address" => address_params}, current_user) do
-    address = Accounts.get_address!(current_user, id)
+    address = UserProfiles.get_address!(current_user, id)
 
-    case Accounts.update_address(address, address_params) do
+    case UserProfiles.update_address(address, address_params) do
       {:ok, address} ->
         conn
         |> put_flash(:info, gettext("Address updated successfully."))
@@ -76,8 +76,8 @@ defmodule VutuvWeb.AddressController do
   end
 
   def delete(conn, %{"id" => id}, current_user) do
-    address = Accounts.get_address!(current_user, id)
-    {:ok, _address} = Accounts.delete_address(address)
+    address = UserProfiles.get_address!(current_user, id)
+    {:ok, _address} = UserProfiles.delete_address(address)
 
     conn
     |> put_flash(:info, gettext("Address deleted successfully."))
