+++ lib/vutuv_web/controllers/api/user_controller.ex
@@ -4,7 +4,7 @@ defmodule VutuvWeb.Api.UserController do
   import VutuvWeb.Api.Authorize
 
   alias Phauxth.Log
-  alias Vutuv.{Accounts, Socials}
+  alias Vutuv.{UserProfiles, Publications}
 
   action_fallback VutuvWeb.Api.FallbackController
 
@@ -17,12 +17,12 @@ defmodule VutuvWeb.Api.UserController do
   end
 
   def index(conn, _, _current_user) do
-    users = Accounts.list_users()
+    users = UserProfiles.list_users()
     render(conn, "index.json", users: users)
   end
 
   def create(conn, %{"user" => user_params}, _current_user) do
-    with {:ok, user} <- Accounts.create_user(user_params) do
+    with {:ok, user} <- UserProfiles.create_user(user_params) do
       Log.info(%Log{user: user.id, message: "user created"})
 
       conn
@@ -34,27 +34,35 @@ defmodule VutuvWeb.Api.UserController do
 
   def show(conn, %{"slug" => slug}, %{"slug" => slug} = current_user) do
     user =
-      Accounts.with_associated_data(current_user, [:email_addresses, :tags, :followers, :leaders])
+      UserProfiles.with_associated_data(current_user, [
+        :email_addresses,
+        :tags,
+        :followers,
+        :leaders
+      ])
 
-    posts = Socials.list_posts(current_user)
+    posts = Publications.list_posts(current_user)
     render(conn, "show.json", user: user, posts: posts)
   end
 
   def show(conn, %{"slug" => slug}, current_user) do
-    user = Accounts.get_user!(%{"slug" => slug})
-    user = Accounts.with_associated_data(user, [:email_addresses, :tags, :followers, :leaders])
-    posts = Socials.list_posts(user, current_user)
+    user = UserProfiles.get_user!(%{"slug" => slug})
+
+    user =
+      UserProfiles.with_associated_data(user, [:email_addresses, :tags, :followers, :leaders])
+
+    posts = Publications.list_posts(user, current_user)
     render(conn, "show.json", user: user, posts: posts)
   end
 
   def update(conn, %{"user" => user_params}, current_user) do
-    with {:ok, user} <- Accounts.update_user(current_user, user_params) do
+    with {:ok, user} <- UserProfiles.update_user(current_user, user_params) do
       render(conn, "show.json", user: user)
     end
   end
 
   def delete(conn, _, user) do
-    {:ok, _user} = Accounts.delete_user(user)
+    {:ok, _user} = UserProfiles.delete_user(user)
     send_resp(conn, :no_content, "")
   end
 end
