+++ lib/vutuv_web/controllers/post_controller.ex
@@ -3,7 +3,7 @@ defmodule VutuvWeb.PostController do
 
   import VutuvWeb.Authorize
 
-  alias Vutuv.{Accounts, Accounts.User, Socials, Socials.Post}
+  alias Vutuv.{Publications, Publications.Post, UserProfiles, UserProfiles.User}
 
   @dialyzer {:nowarn_function, new: 3}
 
@@ -16,23 +16,23 @@ defmodule VutuvWeb.PostController do
   end
 
   def index(conn, %{"user_slug" => slug}, %User{slug: slug} = current_user) do
-    posts = Socials.list_posts(current_user)
+    posts = Publications.list_posts(current_user)
     render(conn, "index.html", posts: posts, user: current_user)
   end
 
   def index(conn, %{"user_slug" => slug}, current_user) do
-    user = Accounts.get_user!(%{"slug" => slug})
-    posts = Socials.list_posts(user, current_user)
+    user = UserProfiles.get_user!(%{"slug" => slug})
+    posts = Publications.list_posts(user, current_user)
     render(conn, "index.html", posts: posts, user: user)
   end
 
   def new(conn, _params, _current_user) do
-    changeset = Socials.change_post(%Post{})
+    changeset = Publications.change_post(%Post{})
     render(conn, "new.html", changeset: changeset)
   end
 
   def create(conn, %{"post" => post_params}, current_user) do
-    case Socials.create_post(current_user, post_params) do
+    case Publications.create_post(current_user, post_params) do
       {:ok, post} ->
         conn
         |> put_flash(:info, gettext("Post created successfully."))
@@ -44,26 +44,26 @@ defmodule VutuvWeb.PostController do
   end
 
   def show(conn, %{"id" => id, "user_slug" => slug}, %User{slug: slug} = current_user) do
-    post = Socials.get_post!(current_user, id)
+    post = Publications.get_post!(current_user, id)
     render(conn, "show.html", post: post, user: current_user)
   end
 
   def show(conn, %{"id" => id, "user_slug" => slug}, current_user) do
-    user = Accounts.get_user!(%{"slug" => slug})
-    post = Socials.get_post!(user, id, current_user)
+    user = UserProfiles.get_user!(%{"slug" => slug})
+    post = Publications.get_post!(user, id, current_user)
     render(conn, "show.html", post: post, user: user)
   end
 
   def edit(conn, %{"id" => id}, current_user) do
-    post = Socials.get_post!(current_user, id)
-    changeset = Socials.change_post(post)
+    post = Publications.get_post!(current_user, id)
+    changeset = Publications.change_post(post)
     render(conn, "edit.html", post: post, changeset: changeset)
   end
 
   def update(conn, %{"id" => id, "post" => post_params}, current_user) do
-    post = Socials.get_post!(current_user, id)
+    post = Publications.get_post!(current_user, id)
 
-    case Socials.update_post(post, post_params) do
+    case Publications.update_post(post, post_params) do
       {:ok, post} ->
         conn
         |> put_flash(:info, gettext("Post updated successfully."))
@@ -75,8 +75,8 @@ defmodule VutuvWeb.PostController do
   end
 
   def delete(conn, %{"id" => id}, current_user) do
-    post = Socials.get_post!(current_user, id)
-    {:ok, _post} = Socials.delete_post(post)
+    post = Publications.get_post!(current_user, id)
+    {:ok, _post} = Publications.delete_post(post)
 
     conn
     |> put_flash(:info, gettext("Post deleted successfully."))
