+++ lib/vutuv_web/controllers/user_controller.ex
@@ -4,7 +4,7 @@ defmodule VutuvWeb.UserController do
   import VutuvWeb.Authorize
 
   alias Phauxth.Log
-  alias Vutuv.{Accounts, Accounts.User, Devices, Socials}
+  alias Vutuv.{UserProfiles, UserProfiles.User, Devices, Publications}
   alias VutuvWeb.EmailAddressController
 
   @dialyzer {:nowarn_function, new: 3}
@@ -18,7 +18,7 @@ defmodule VutuvWeb.UserController do
   end
 
   def index(conn, params, _current_user) do
-    page = Accounts.paginate_users(params)
+    page = UserProfiles.paginate_users(params)
     render(conn, "index.html", users: page.entries, page: page)
   end
 
@@ -27,7 +27,7 @@ defmodule VutuvWeb.UserController do
   end
 
   def new(conn, _, _current_user) do
-    changeset = Accounts.change_user(%User{})
+    changeset = UserProfiles.change_user(%User{})
     render(conn, "new.html", changeset: changeset)
   end
 
@@ -35,7 +35,7 @@ defmodule VutuvWeb.UserController do
     user_params =
       conn |> get_req_header("accept-language") |> add_accept_language_to_params(user_params)
 
-    case Accounts.create_user(user_params) do
+    case UserProfiles.create_user(user_params) do
       {:ok, user} ->
         Log.info(%Log{user: user.id, message: "user created"})
         EmailAddressController.verify_email(conn, user_params, "confirm your account", true)
@@ -51,26 +51,34 @@ defmodule VutuvWeb.UserController do
 
   def show(conn, %{"slug" => slug}, %{"slug" => slug} = current_user) do
     user =
-      Accounts.with_associated_data(current_user, [:email_addresses, :tags, :followers, :leaders])
+      UserProfiles.with_associated_data(current_user, [
+        :email_addresses,
+        :tags,
+        :followers,
+        :leaders
+      ])
 
-    posts = Socials.list_posts(current_user)
+    posts = Publications.list_posts(current_user)
     render(conn, "show.html", user: user, posts: posts)
   end
 
   def show(conn, %{"slug" => slug}, current_user) do
-    user = Accounts.get_user!(%{"slug" => slug})
-    user = Accounts.with_associated_data(user, [:email_addresses, :tags, :followers, :leaders])
-    posts = Socials.list_posts(user, current_user)
+    user = UserProfiles.get_user!(%{"slug" => slug})
+
+    user =
+      UserProfiles.with_associated_data(user, [:email_addresses, :tags, :followers, :leaders])
+
+    posts = Publications.list_posts(user, current_user)
     render(conn, "show.html", user: user, posts: posts)
   end
 
   def edit(conn, _, user) do
-    changeset = Accounts.change_user(user)
+    changeset = UserProfiles.change_user(user)
     render(conn, "edit.html", user: user, changeset: changeset)
   end
 
   def update(conn, %{"user" => user_params}, user) do
-    case Accounts.update_user(user, user_params) do
+    case UserProfiles.update_user(user, user_params) do
       {:ok, user} ->
         conn
         |> put_flash(:info, gettext("User updated successfully."))
@@ -82,7 +90,7 @@ defmodule VutuvWeb.UserController do
   end
 
   def delete(conn, _, user) do
-    {:ok, _user} = Accounts.delete_user(user)
+    {:ok, _user} = UserProfiles.delete_user(user)
 
     conn
     |> delete_session(:phauxth_session_id)
