+++ test/vutuv/user_profiles/locale_test.exs
@@ -1,7 +1,7 @@
-defmodule Vutuv.Accounts.LocaleTest do
+defmodule Vutuv.UserProfiles.LocaleTest do
   use ExUnit.Case
 
-  alias Vutuv.Accounts.Locale
+  alias Vutuv.UserProfiles.Locale
 
   test "locale with top quality is selected" do
     al = "en-ca,en;q=0.8,en-us;q=0.6,de-de;q=0.4,de;q=0.2"
