+++ test/vutuv_web/controllers/follower_controller_test.exs
@@ -3,12 +3,12 @@ defmodule VutuvWeb.FollowerControllerTest do
 
   import Vutuv.Factory
 
-  alias Vutuv.Accounts
+  alias Vutuv.UserProfiles
 
   test "lists all of a user's followers", %{conn: conn} do
     user = insert(:user)
     followers = insert_list(12, :user)
-    for follower <- followers, do: Accounts.add_leaders(follower, [user.id])
+    for follower <- followers, do: UserProfiles.add_leaders(follower, [user.id])
     conn = get(conn, Routes.user_follower_path(conn, :index, user))
     assert html_response(conn, 200) =~ "Followers"
   end
