+++ test/vutuv_web/controllers/leader_controller_test.exs
@@ -3,12 +3,12 @@ defmodule VutuvWeb.LeaderControllerTest do
 
   import Vutuv.Factory
 
-  alias Vutuv.Accounts
+  alias Vutuv.UserProfiles
 
   test "lists all of a user's leaders (following)", %{conn: conn} do
     user = insert(:user)
     leader_ids = Enum.map(insert_list(12, :user), & &1.id)
-    Accounts.add_leaders(user, leader_ids)
+    UserProfiles.add_leaders(user, leader_ids)
     conn = get(conn, Routes.user_leader_path(conn, :index, user))
     assert html_response(conn, 200) =~ "Following"
   end
