+++ test/vutuv_web/controllers/post_controller_test.exs
@@ -4,7 +4,7 @@ defmodule VutuvWeb.PostControllerTest do
   import Vutuv.Factory
   import VutuvWeb.AuthTestHelpers
 
-  alias Vutuv.{Accounts, Socials}
+  alias Vutuv.{UserProfiles, Publications}
 
   @create_attrs %{
     body: Faker.Company.bs(),
@@ -35,7 +35,7 @@ defmodule VutuvWeb.PostControllerTest do
       post_1 = insert(:post, %{user: user, visibility_level: "public"})
       post_2 = insert(:post, %{user: user, visibility_level: "followers"})
       other = add_user("froderick@example.com")
-      {:ok, _user} = Accounts.add_leaders(other, [user.id])
+      {:ok, _user} = UserProfiles.add_leaders(other, [user.id])
       conn = conn |> add_session(other) |> send_resp(:ok, "/")
       conn = get(conn, Routes.user_post_path(conn, :index, user))
       response = html_response(conn, 200)
@@ -67,7 +67,7 @@ defmodule VutuvWeb.PostControllerTest do
       end
 
       other = add_user("froderick@example.com")
-      {:ok, _user} = Accounts.add_leaders(other, [user.id])
+      {:ok, _user} = UserProfiles.add_leaders(other, [user.id])
       conn = conn |> add_session(other) |> send_resp(:ok, "/")
       conn = get(conn, Routes.user_post_path(conn, :show, user, post))
       assert html_response(conn, 200) =~ dirty_escape(post.title)
@@ -122,7 +122,7 @@ defmodule VutuvWeb.PostControllerTest do
       conn = put(conn, Routes.user_post_path(conn, :update, user, post), post: @update_attrs)
       assert redirected_to(conn) == Routes.user_post_path(conn, :show, user, post)
       assert get_flash(conn, :info) =~ "updated successfully"
-      post = Socials.get_post!(user, post.id)
+      post = Publications.get_post!(user, post.id)
       assert post.visibility_level == "public"
     end
 
@@ -145,7 +145,7 @@ defmodule VutuvWeb.PostControllerTest do
       conn = delete(conn, Routes.user_post_path(conn, :delete, user, post))
       assert redirected_to(conn) == Routes.user_post_path(conn, :index, user)
       assert get_flash(conn, :info) =~ "deleted successfully"
-      assert_raise Ecto.NoResultsError, fn -> Socials.get_post!(user, post.id) end
+      assert_raise Ecto.NoResultsError, fn -> Publications.get_post!(user, post.id) end
     end
 
     test "cannot delete another user's post", %{conn: conn, user: user} do
@@ -156,7 +156,7 @@ defmodule VutuvWeb.PostControllerTest do
         delete(conn, Routes.user_post_path(conn, :delete, user, post))
       end
 
-      assert Socials.get_post!(other, post.id)
+      assert Publications.get_post!(other, post.id)
     end
   end
